+++ src/DynamoDb/DynamoDbClient.php
@@ -21,6 +21,8 @@ use Aws\RetryMiddleware;
  * @method \GuzzleHttp\Promise\Promise deleteItemAsync(array $args = [])
  * @method \Aws\Result deleteTable(array $args = [])
  * @method \GuzzleHttp\Promise\Promise deleteTableAsync(array $args = [])
+ * @method \Aws\Result describeLimits(array $args = [])
+ * @method \GuzzleHttp\Promise\Promise describeLimitsAsync(array $args = [])
  * @method \Aws\Result describeTable(array $args = [])
  * @method \GuzzleHttp\Promise\Promise describeTableAsync(array $args = [])
  * @method \Aws\Result getItem(array $args = [])